requirements.txt
# FastAPI and ASGI
fastapi==0.115.6
uvicorn[standard]==0.34.0
uvloop==0.21.0
httptools==0.6.4

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis==5.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# HTTP client for orchestrator RPC
httpx==0.28.1

# Background tasks
apscheduler==3.10.4

# Settings management
pydantic-settings==2.6.1
email-validator==2.2.0

# WebSocket support (included in uvicorn[standard])

# Utilities
python-multipart==0.0.18
python-dotenv==1.0.1

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
